Initialize network's vertex attributes

Label vertex attributes of a conditional graphical model's output.

adj2igraph(YY, XY = NULL, info, weighted = F, dropnull = TRUE)

Arguments

YY

adjacency matrix encoding y--y edges of conditional graphical model. If a logical matrix, TRUE denotes the existence of an edge. Format should conform to igraph::graph_from_adjacency_matrix input.

XY

adjacency matrix encoding x--y edges of conditional graphical model. If NULL, then a network encoding just y--y edges is made and info should just contain attributes for (y(1), ..., y(Q)).

info

data.frame encoding the attributes of the vertices in the order of (x(1), ..., x(p), y(1), ..., y(q)) if both YY and XY are specified. If just YY is specified, then info should just contain attributes for (y(1), ..., y(Q)). See details for required attributes to include and optional attributes.

weighted

logical indicating whether the values of YY and XY should indicate weights of the vertices. Defaults to FALSE.

dropnull

logical indicating to remove vertices with degree zero from the graph. Defaults to TRUE.