Ensemble networks through boostrapping.

Fits spacemap model to B bootstrapped data replicates to create an ensemble network.

bootEnsemble(Y.m, X.m, tune, boot = TRUE, B = 1000L, seed = 55139,
  p0 = 1, iter = 3, tol = 1e-04, cd_iter = 1e+09, refitRidge = 0.1,
  iscale = FALSE)

Arguments

Y.m
Numeric matrix \((N \times Q)\) containing N iid samples of the response vector \(\textbf{y}\).

X.m
Numeric matrix \((N \times P)\) containing N iid samples of the predictor vector \(\textbf{x}\).

tune
Data.frame with three columns requiring names to be the tuning penalties slasso, rlasso, rgroup that are input to spacemap. Each row is a tuning penalty set that indexes an ensemble network of spacemap fits to B boostrap replicates of Y.m, X.m.

boot
Logical. Default is boot=TRUE and implies sampling with replacement (i.e. boostrap). boot=FALSE will resample without replacement.

B
Positive integer denoting the number of model fits making up the ensemble. Default is 1000, but can be lowered to save time.

seed
Positive integer allowing user to set the random seed for reproducing the network ensemble.

p0
Positive numeric not exceeding the default of 1, which represents the proportion of the original samples that will be down-sampled for each model fitting iteration of the ensemble.

iter
Positive integer specifying the number of iterations for estimating \(\sigma^{ii}\) and partial correlations. Defaults to 3. When sig=NULL and/or (weight=NULL or weight=2), then iter \(\ge 2\).

tol
Positive numeric value specifying the convergence tolerance of the coordinate descent algorithm; in other words, it is criterion that stops parameter estimation when no parameter changes value exceeding tol between iterations. tol defaults to 1e-6, but may be lowered (e.g. 1e-4) to speed up network learning.

cd_iter
Positive integer specifiying the maximum number of parameter updates allowed before reporting the algorithm as having failed to converge. Default may need to be increased for inferring very large-scale networks (i.e. \(p,q > 1000\)).

refitRidge
Positive numeric value specifying a ridge penalty, which is applied when refitting the model to the data after having selected the network with spacemap. Defaults to 0.1, but can be set to zero for ordinary least squares refit.

iscale
Logical indicating to standardize the input data internally. Defaults to TRUE if just running spacemap. If running crossValidation or bootEnsemble, user should set to FALSE.

Value

a list of length B, where the bth element contains the refit of spacemap model. The model output is identical to spacemap, but the partial correlations matrix ParCor and the regression coefficients matrix Gamma are stored in a sparse matrix format, (dgCMatrix format), from the Matrix package. If the bth model is not sufficiently sparse, a character vector will be returned as "Not Sparse". If the bth model did not converge, a character vector will be returned as "No Convergence".

Examples

#Load simulation
data(sim1)
#Boostrap Ensemble (B = 10) for illustration only
tune <- data.frame(slasso = 70, rlasso = 28.8, rgroup = 14)
out <- bootEnsemble(Y.m = sim1$Y, X.m = sim1$X, tune = tune, B = 10, iter = 3,
                    cd_iter = 1e7, tol = 1e-6, refitRidge = 0)
bv <- bootVote(fits = out[[1]], P = ncol(sim1$X), Q = ncol(sim1$Y))