Cross validation (CV.Vote) for spacemap and space models.

Selects and returns best-tuned model under CV.Vote.

crossValidation(data, fold, trainIds, testIds, method = c("spacemap",
  "space"), tuneGrid, refit = TRUE, ...)

Arguments

data

list containing named elements of at least X and Y. Element XY is required when method = "space". X is an (N x P) matrix containing N samples and P predictor variables. Y is an (N X Q) matrix containing N samples and Q response variables. XY is an (N X (P + Q)) matrix generated by cbind(X,Y).

fold

integer reporting the number of cross validation folds (e.g. 10)

trainIds

list (named optional) of size fold with the ith element containing the sample indices of corresponding to the ith training set.

testIds

list (named optional) of size fold with the ith element containing the sample indices of corresponding to the ith test set.

method

character vector reporting to fit either Spacemap or Space.

tuneGrid

data.frame named with columns slasso, rlasso, rgroup when method = "spacemap". Each row in the data.frame corresponds to a tuning parameter set that is input into Spacemap. So the size of the tuning grid is nrow(tuneGrid). When method = "space", the previous description still applies, but is a data.frame with the only column being slasso.

refit

logical telling to refit the model after convergence with either ordinary least squares or ridge regression. Must specify refitRidge = 0.0 as an additional argument to indicate OLS, otherwiserefitRidge = 0.001 as a default.

...

Additional arguments for Spacemap and Space. For both methods specify values for iter, cd_iter, tol, refitRidge. For Space, sridge, Xindex, Yindex should be specified. Set sridge = 0.0 unless elastic net penalty is desired. Xindex are the indices of the X variables in the XY element of data. Yindex are the indices of the Y variables in the XY element of data.

Value

List with elements (i) cvFits which contains the cv-selected fitted models according to cv-selection type. tuneGrid the selected tuning parameter set. minIndices the minimum CV-score index of the input tuning grid. metricScores A list of Residual sums of squares; total degrees of freedom, Partial Correlation degrees of freedom, Gamma degrees of freedom, and the final maximum difference between parameter estimates useful for convergence diagnostics. Each of these is laid out in a matrix of size (tuning grid X number of CV folds).

Examples


## Not run: ------------------------------------
# 
# 
# #load data
# data(sim1)
# dat <- sim1[c("XY", "X", "Y", "Xindex", "Yindex")]
# N <- nrow(dat$X)
# P <- ncol(dat$X)
# Q <- ncol(dat$Y)
# #set up parallel environment
# dopar <- TRUE
# if (dopar) { 
#  suppressPackageStartupMessages(library(doParallel))
#  suppressPackageStartupMessages(library(parallel))
#  ncores <- detectCores()  - 1
#  cl <- makeCluster(ncores)
#  registerDoParallel(cl)
# }
# 
# #set up training and test sets
# K <- 10L
# set.seed(265616L)
# allSets <- seq_len(N)
# library(caret)
# testSets <- caret::createFolds(allSets, k = K)
# trainSets <- lapply(testSets, function(s) setdiff(allSets, s))
# 
# #create tuning grid for spacemap
# tmap <- expand.grid(slasso = c(65,70,73),
#                    rlasso = c(27,28.8, 30),
#                    rgroup = c(11, 12.38, 14))
# #cross validation of spacemap
# out <- spacemap::crossValidation(dat = dat, fold = K,
#                                 trainIds = trainSets,  testIds = testSets, 
#                                 method = "spacemap", tuneGrid = tmap, 
#                                 iter = 3, tol = 1e-6, cd_iter = 1e7, 
#                                 sridge = 0, refitRidge = 0)
## ---------------------------------------------