Network analysis toolkit

Downstream analysis of spaceMap networks

neta(net, bdeg, info, go2eg, cis_window = 2e+06, lmods = NULL,
  min_module_size = 15, mod_prefix = "P", fdr_thresh = 0.05, min_go = 5)

Arguments

net
list containing two adjacency matrices, named yy and xy, encoding the response--response interactions and the predictor--response interactions, respectively.

bdeg
list containing two matrices, xy and yy, where each row represents a bootstrap relicate of the degree distribution and the columns contain the degree across all x nodes (in the case of xy ) or all y nodes (in the case of yy).

info
data.frame containing several columns that describe the attributes of the nodes in the network. The columns are currently required to the following variables, but they can be missing values for all variables but the ID and type of node. The variables are: id (a unique ID for the gene node); chr (chromosome location); start (starting coordinate on the chromosome); end (ending coordinate); cyto (cyotband abbreviation, e.g. 5q34); cytoscape (a gene symbol); strand (the +/- strand); type (either a "x" or "y" corresponding to predictor or response nodes).

go2eg
named list where the names denote a biological process (e.g. GO ID) and the elements of the list is a vector of members belonging to the biological process. The list ought to be non-redundant in names.

cis_window
numeric denoting the half length of the genomic interval used for calling a response node (i.e. of type y) as cis regulated by a predictor node (i.e. of type x)..

lmods
named list of modules in the network. When lmdods=NULL, the Girvan-Newman edge betweenness algorithm is used to find network modules.

min_module_size
integer defining the minimum size a module must be to be tested for enrichment.

mod_prefix
character to prefix module ID's.

fdr_thresh
numeric between 0 and 1 to denote the FDR threshold in the Bonferroni-Hochberg correction.

min_go
integer defining the minimum number of nodes required to be to call a significant enrichment of a biological process based on the hypergeometric test.

Value

list

ig The igraph object encoding the network and its attributes.

xhubtab data.frame reporting prioritized predictor hubs and their attributes.

modenrichtab data.frame reporting which modules were enriched with biological processes and what they were. If the module contains predictor nodes (e.g. CNA) with at least min_go edges to other module members it is also reported for the module.

lmodtab list of data.frames corresponding to each module of size at least min_module_size. Each data.frame reports the nodes of the module and other information such as degree and which nodes participate in specific enrichment of biological processes.

goneighbors data.frame where each row corresponds to a predictor hub. The predictor hub induces a neighborhood of response nodes with which it has an edge. A certain proportion of response node neighbors share the same biological processes and this is reported.