Prioritize networks hubs with consistently high degree.

Rank hubs by degree either with the final network or according to how consistently the degree is high across bootstrap replicates.

rankHub(ig, bdeg = NULL, level = c("x", "y"))

Arguments

ig
An igraph network object output from adj2igraph

bdeg
If NULL, the ranking of hubs nodes is done by their highest degree in the ig network within their respective node level = "y" or level = "x".

If non-null, List output from list element of bootVote. Element bdeg$yy[b,] is an integer vector representing the degree distribution for the \(b\)th bootstrap replicate across the y nodes. Similarly, element bdeg$xy[b,] is an integer vector representing the degree distribution for the \(b\)th bootstrap replicate across the x nodes. If `bdeg` is available, the hubs will be prioritized according to their mean rank of degree across the \(B\) bootstrap replicates. Highly ranked hubs consistently have a larger degree than other nodes across the bootstrap replicates.

level
Character value either 'y' or 'x' (defaults to 'x') specifying to rank hubs that are x or y nodes. Ranking levels x and y together is not supported.

Value

The igraph object ig updated with a 'rank_hub' attribute containing each nodes' respective rank. If bdeg is non-null the mean degree rank across bootstrap replicates is labeled with attribute 'mean_rank_hub', and its corresponding standard deviation is labeled as attribute 'sd_rank_hub'.