Report top hub nodes with attributes.

Organize x and y hub analysis into a publish-ready format.

reportHubs(ig, top = 10, level = c("x", "y"), unit = "Mb",
  unit_base = 1e+06, sig = 2)

Arguments

ig
An igraph network object returned from either rankHub or cisTrans.

top
An integer specifying the number of top-degree hubs to report.

level
Character specifying nodes of level x or y, but not both. Defaults to x.

unit
If gene coordinates for x-hubs exist, then this will label the chosen genome location scale in the table. Defaults to mega-base (Mb).

unit_base
Numeric divisor of the genome coordinates and should correspond to unit label. Defaults to 1e6 for mega-base.

sig
Integer for controlling precision of genome coordinates.

Value

A table of top ranked node hubs with attributes deemed relevant to report. See vignette on network analysis for examples.

See also

rankHub, cisTrans, adj2igraph, xHubEnrich, modEnrich