Report joint performance of (x-y,y-y) edge detection

Assesses the network learning performance of spaceMap (or space) against a known truth by reporting power and false discovery rate.

reportJointPerf(fit, truth, aszero = 1e-06, verbose = TRUE)

Arguments

fit
A list of numeric matrices named 'xy' and 'yy' encoding the fitted model parameter estimates (or the true adjacency matrices).

truth
A list of numeric matrices named 'xy' and 'yy' encoding the true model parameters (or an adjacency matrix).

aszero
A numeric value specifying the point at which a parameter estimate should be effectively considered of zero value.

verbose
A logical value indicating whether to report the performance to the console.

Value

A numeric, named vector

power Joint power on (x-y,y-y)

fdr Joint false discovery rate (FDR) on (x-y,y-y)

powerXY Power of x-y edges

fdrXY FDR of x-y edges

powerYY Power of y-y edges

fdrYY FDR of y-y edges

tp True positives of (x-y,y-y)

fn False negatives of (x-y,y-y)

fp False positives of (x-y,y-y)

tpXY True positives of x-y

fnXY False negatives of x-y

fpXY False positives of x-y

tpYY True positives of y-y

fnYY False negatives of y-y

fpYY False positives of y-y

See also

reportPerf