Report Model Performance

Assesses the performance of space, remMap, and spacemap models against a known truth by reporting power and false discovery rate.

reportPerf(fitParCor, trueParCor, YY = TRUE, tol = 1e-06, verbose = TRUE)

Arguments

fitParCor
A numeric matrix encoding the fitted model parameter estimates.

trueParCor
A numeric matrix encoding the true model parameters.

YY
A logical value indicating that fitParCor estimates have the same structure as the Space model. If YY=TRUE, then the fitted and the truth inputs must be symmetric matrices.

tol
A numeric value specifying a lower bound on the tolerance of a non-zero parameter.

verbose
A logical value indicating whether to report the performance to the console.

Value

A numeric, named vector where: (X,Y) is the joint edge performance of X-Y and Y-Y edges.

powerJoint power on (X,Y)

fdrJoint false discovery rate (FDR) on (X,Y)

powerXYPower of X-Y edges

fdrXYFDR of X-Y edges

powerYYPower of Y-Y edges

fdrYYFDR of Y-Y edges

tpTrue positives of (X,Y)

fnFalse negatives of (X,Y)

fpFalse positives of (X,Y)

tpXYTrue positives of X-Y

fnXYFalse negatives of X-Y

fpXYFalse positives of X-Y

tpYYTrue positives of Y-Y

fnYYFalse negatives of Y-Y

fpYYFalse positives of Y-Y