Fit spacemap model

Estimate partial correlations between response variables (Y) and perform model selection of predictors (X) based on the spacemap model fitted to multivariate data (Y,X).

spacemap(Y.m, X.m, slasso, sridge = 0, rlasso, rgroup, sig = NULL,
  weight = NULL, remWeight = NULL, iter = 3, tol = 1e-06,
  cd_iter = 10000000L, verbose = FALSE)

Arguments

Y.m

A numeric matrix (N x Q) representing the response of Q variables on N independent observations.

X.m

A numeric matrix (N X P) representing P predictor variables on N independent observations.

slasso

A numeric value, the Space-lasso penalty parameter (l_1 norm).

sridge

A numeric value (optional). If not specified, lasso regression is used in the Joint Sparse Regression Model (JSRM). Otherwise, elastic net regression is used in JSRM and sridge serves as the squared l_2 norm penalty parameter.

rlasso

The MAP-lasso penalty.

rgroup

The MAP-group penalty.

sig

A numeric vector with length being the same as the number of columns of Y.m. It is the vector of \(\sigma^{ii}\) (the diagonal of the inverse covariance matrix). If not specified, s^ii will be estimated during the model fitting with initial values rep(1,p) and the iterations of the model fitting (iter) \(\ge 2\). Note, the scale of sig does not matter.

weight

A numeric value or vector (Q x 1) specifying the weights or the type of weights used for each regression in JSRM. The default value is NULL, which means all regressions will be weighted equally in the joint model. If weight=1, residue variances will be used for weights. If weight=2, the estimated degree of each variable will be used for weights. Otherwise, it should be a positive numeric vector, whose length is equal to the number of columns of Y.m.

remWeight

optional weighting scheme on [fill in details].

iter

An integer specifying the iterations in JSRM for estimating \(\sigma^{ii}\) and partial correlations. When sig=NULL and/or (weight=NULL or weight=2), then iter \(\ge 2\).

Value

A list containing

ParCor The estimated partial correlation matrix.

sig.fit The estimated diagonal \(\sigma^{ii}\).

Gamma The estimated coefficients of Y on X.

rss The residual sums of squares from the model fit.

deltaMax The maximum change in parameter values between the penultimate and ultimate iteration

Examples

data(sim1)
net <- spacemap(Y.m = sim1$Y, X.m = sim1$X, slasso = 70, rlasso = 28.8, rgroup = 12.38)