Visualize cross validation metrics

Diagnose the quality of the tuning grid used in cvVote through plotting CV scores and network size against the tuning penalties.

tuneVis(cvOut, testSetLen, tuneParam1, tuneParam1Name = c("lam1", "lam2",
  "lam3"), tuneParam2 = NULL)

Arguments

cvOut
Output from cvVote.

testSetLen
Integer vector specifying length of each test set associated with cvOut.

tuneParam1
Numeric vector specifying the tuning penalty placed on the x-axis.

tuneParam1Name
Character vector labelling the X-axis with the specified tuning parameter.

tuneParam2
Numeric vector specifying the tuning penalty placed on the x-axis. Defaults to NULL. If non-null, the output is a heatmap where CV scores are colored intensities on a 2-D grid of tuneParam1,tuneParam2.

Value

If tuneParam2 == NULL Return a list of 4 ggplot2 objects.

Scatter plot of log (CV score) vs. tuneParam1.

Scatter plot of "Avg. Total # of Edges" vs. tuneParam1.

Scatter plot of "Avg. # of Edges (Y -- Y)" vs. tuneParam1.

Scatter plot of "Avg. # of Edges (x -> Y)" vs. tuneParam1.

If tuneParam2 != NULL Returns a heatmap (see ggplot2::geom_tile) where CV scores are colored intensities on a 2-D grid of tuneParam1,tuneParam2.

See also

cvVote