Ensemble networks through boostrapping.

Fits spacemap model to B bootstrapped data replicates to create an ensemble network.

bootEnsemble(Y, X = NULL, tune, method = c("spacemap", "space"),
  boot = TRUE, B = 1000L, iscale = TRUE, aszero = 1e-06,
  seed = sample.int(n = 1e+06, size = 1), p0 = 1, ...)

Arguments

Y
Numeric matrix \((N \times Q)\) containing N iid samples of the response vector \(\textbf{y}\).

X
Numeric matrix \((N \times P)\) containing N iid samples of the predictor vector \(\textbf{x}\).

tune
List with names being the tuning penalties lam1, lam2, lam3 and each element being a single numeric value that is input to spacemap. One might consider the output from cvVote (i.e. list minTune) as input for this parameter.

boot
Logical. Default is boot=TRUE and implies sampling with replacement (i.e. boostrap). boot=FALSE will resample without replacement.

B
Positive integer denoting the number of model fits making up the ensemble. Default is 1000, but can be lowered to save time.

iscale
Logical indicating to standardize the whole input data. Defaults to TRUE. See base::scale(x, center = TRUE, scale = TRUE) for details of standardization.

aszero
Positive numeric value (defaults to 1e-6) indicating at what point to consider extremely small parameter estimates of \(\Gamma\) and \(\rho\) as zero.

seed
Positive integer allowing user to set the random seed for reproducing the network ensemble.

p0
Positive numeric not exceeding the default of 1, which represents the proportion of the original samples that will be down-sampled for each model fitting iteration of the ensemble.

Value

A list up to length B of convergent and sparse model fits from either spacemap or space.joint. The list object should typically not be modified by the user, but passed on to the bootVote function. Model fits that are not convergent or produce networks that are not sparse are not reported.

See also

bootVote, cvVote

Examples


#Load simulation
data(sim1)

#Boostrap Ensemble (B = 10) for illustration only
tune <- data.frame(lam1 = 70, lam2 = 28, lam3 = 17.5)
ens <- bootEnsemble(Y = sim1$Y, X = sim1$X, tune = tune,
                    method = "spacemap", B = 10)

#> Loading required package: foreach

#> foreach: simple, scalable parallel programming from Revolution Analytics
#> Use Revolution R for scalability, fault tolerance and more.
#> http://www.revolutionanalytics.com

#> Loading required package: rngtools

#> Loading required package: pkgmaker

#> Loading required package: registry

#> 
#> Attaching package: ‘pkgmaker’

#> The following object is masked from ‘package:base’:
#> 
#>     isNamespaceLoaded

#> Warning: executing %dopar% sequentially: no parallel backend registered