Model aggregation through Boot.Vote

Aggregate boostrap replicates of spacemap into a final Boot.Vote model.

bootVote(fits, P, Q, vote_thresh = 0.5, btol = 0)

Arguments

fits
List of fitted spacemap models returned from bootEnsemble.

P
Integer for the number of predictor variables

Q
Integer for the number of response variables

vote_thresh
Positive numeric threshold for the minimum proportion of bootstrap replicate model fits with a particular edge such that the edge is included in the Boot.Vote model.

btol
The numeric lower bound that indicates parameter estimates equal to or less than btol ought to not be considered as an edge in a model fit. Do not change from zero if used in tandem with bootEnsemble.

Value

List containing two lists. First list is bv, which encodes the edges in two logical adjacency matrices.

ParCor Logical matrix where TRUE for the (q,l) off-diagonals element indicate an edge between the qth and lth response variables.

Gamma Logical matrix where TRUE for the (p,q) element indicate an edge between the pth predictor and qth response variable.

Second list is bc, which stores several additional statistics on the ensemble network fits.

ParCor Integer matrix containing the y--y edge selection frequency out of B replicates.

Gamma Integer matrix containing the x--y edge selection frequency out of B replicates.

dfParCor Integer vector containing the total number of y--y edges for each fit.

dfGamma Integer vector containing the total number of x--y edges for each fit.

nuniq Logical vector where TRUE indicates the model fit converged and was sparse and FALSE otherwise.

Examples

#Load simulation
data(sim1)
#Boostrap Ensemble (B = 10) for illustration only
tune <- data.frame(slasso = 70, rlasso = 28.8, rgroup = 14)
out <- bootEnsemble(Y.m = sim1$Y, X.m = sim1$X, tune = tune, B = 10, iter = 3,
                    cd_iter = 1e7, tol = 1e-6, refitRidge = 0)
bv <- bootVote(fits = out[[1]], P = ncol(sim1$X), Q = ncol(sim1$Y))